Isolate keyring instances based on tenant-id and client-id
Valid authentication cache for a certain tenant-id / client-id pair should
not be re-used when a user specifies a different pair, whether as an update
to the same remote repository settings or the addition of a new remote.
Likewise, authentication cache retrieved for a newly configured tenant-id /
client-id pair should not clobber the cache of a previously used pair.

Signed-off-by: Rahul Rameshbabu <sergeantsagara@protonmail.com>